requirements.txt
fastapi
uvicorn
google-genai
python-multipart
python-dotenv
aiofiles
slowapi
markdownjinja2
